Rename UpdateApplicationFilesDto import to match updated file name
