+++ src/application/application.controller.ts
@@ -25,7 +25,7 @@ import {
   ApiConsumes,
 } from '@nestjs/swagger';
 import { UpdateApplicationDto } from './dto/update-application.dto';
-import { UpdateApplicationFilesDto } from './dto/update-application-file.dto';
+import { UpdateApplicationFilesDto } from './dto/update-application-files.dto';
 
 const apiBodyOptions: ApiBodyOptions = {
   schema: {
